many: move new environment code to osutil
There's a lot of churn but that's unavoidable in such a move.
